refactor: gateway template, function
- drop local AuthMode, use cluster.AuthenticationMode instead
- drop old function detectClusterAuthModedetectClusterAuthMode, use
 cluster.GetClusterAuthenticationMode instead
 - rename two existing templates and  use const to refer them
