+++ internal/controller/services/gateway/gateway_auth_actions.go
@@ -11,6 +11,7 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
 	serviceApi "github.com/opendatahub-io/opendatahub-operator/v2/api/services/v1alpha1"
+	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/cluster"
 	odhtypes "github.com/opendatahub-io/opendatahub-operator/v2/pkg/controller/types"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/resources"
 )
@@ -31,7 +32,7 @@ func createKubeAuthProxyInfrastructure(ctx context.Context, rr *odhtypes.Reconci
 		return fmt.Errorf("failed to resolve domain: %w", err)
 	}
 
-	authMode, err := detectClusterAuthMode(ctx, rr)
+	authMode, err := cluster.GetClusterAuthenticationMode(ctx, rr.Client)
 	if err != nil {
 		return fmt.Errorf("failed to detect cluster authentication mode: %w", err)
 	}
@@ -45,7 +46,7 @@ func createKubeAuthProxyInfrastructure(ctx context.Context, rr *odhtypes.Reconci
 	}
 
 	var oidcConfig *serviceApi.OIDCConfig
-	if authMode == AuthModeOIDC {
+	if authMode == cluster.AuthModeOIDC {
 		oidcConfig = gatewayConfig.Spec.OIDC
 	}
 
@@ -59,7 +60,7 @@ func createKubeAuthProxyInfrastructure(ctx context.Context, rr *odhtypes.Reconci
 		return fmt.Errorf("failed to deploy auth proxy: %w", err)
 	}
 
-	if authMode == AuthModeIntegratedOAuth {
+	if authMode == cluster.AuthModeIntegratedOAuth {
 		if err := createOAuthClient(ctx, rr, clientSecret); err != nil {
 			return fmt.Errorf("failed to create OAuth client: %w", err)
 		}
@@ -95,7 +96,7 @@ func createEnvoyFilter(ctx context.Context, rr *odhtypes.ReconciliationRequest) 
 	authTimeout := getGatewayAuthTimeout(gatewayConfig)
 
 	// using yaml templates due to complexity of k8s api struct for envoy filter
-	yamlContent, err := gatewayResources.ReadFile("resources/envoyfilter-authn.yaml")
+	yamlContent, err := gatewayResources.ReadFile(envoyFilterTemplate)
 	if err != nil {
 		return fmt.Errorf("failed to read EnvoyFilter template: %w", err)
 	}
