+++ internal/controller/services/gateway/gateway_support.go
@@ -9,7 +9,6 @@ import (
 	"os"
 	"strings"
 
-	configv1 "github.com/openshift/api/config/v1"
 	oauthv1 "github.com/openshift/api/oauth/v1"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -32,8 +31,10 @@ import (
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/resources"
 )
 
-// AuthMode represents different authentication modes supported by the gateway.
-type AuthMode string
+const (
+	ServiceName        = serviceApi.GatewayServiceName
+	ReadyConditionType = serviceApi.GatewayConfigKind + status.ReadySuffix
+)
 
 const (
 	// Gateway infrastructure constants.
@@ -74,10 +75,16 @@ const (
 
 	// OAuth2 proxy cookie name - used in both proxy args and EnvoyFilter Lua filter.
 	OAuth2ProxyCookieName = "_oauth2_proxy"
+)
 
-	AuthModeIntegratedOAuth AuthMode = "IntegratedOAuth"
-	AuthModeOIDC            AuthMode = "OIDC"
-	AuthModeNone            AuthMode = "None"
+const (
+	envoyFilterTemplate     = "resources/envoyfilter-authn.tmpl.yaml"
+	destinationRuleTemplate = "resources/kube-auth-proxy-destinationrule-tls.tmpl.yaml"
+	// not in use yet, TODO comment this out.
+	// kubeAuthProxyDeploymentOidcTemplate  = "resources/kube-auth-proxy-oidc-deployment.tmpl.yaml"
+	// kubeAuthProxyDeploymentOauthTemplate = "resources/kube-auth-proxy-oauth-deployment.tmpl.yaml"
+	// kubeAuthProxyServiceTemplate         = "resources/kube-auth-proxy-svc.tmpl.yaml"
+	// kubeAuthProxyHTTPRouteTemplate       = "resources/kube-auth-proxy-httproute.tmpl.yaml".
 )
 
 var (
@@ -350,29 +357,8 @@ func createGateway(rr *odhtypes.ReconciliationRequest, certSecretName string, do
 	return rr.AddResources(gateway)
 }
 
-// detectClusterAuthMode determines the authentication mode from cluster configuration.
-func detectClusterAuthMode(ctx context.Context, rr *odhtypes.ReconciliationRequest) (AuthMode, error) {
-	auth := &configv1.Authentication{}
-	err := rr.Client.Get(ctx, types.NamespacedName{Name: "cluster"}, auth)
-	if err != nil {
-		return "", fmt.Errorf("failed to get cluster authentication config: %w", err)
-	}
-
-	switch auth.Spec.Type {
-	case "OIDC":
-		return AuthModeOIDC, nil
-	case "IntegratedOAuth", "":
-		// empty string is equivalent to IntegratedOAuth (default)
-		return AuthModeIntegratedOAuth, nil
-	case "None":
-		return AuthModeNone, nil
-	default:
-		return AuthModeIntegratedOAuth, nil
-	}
-}
-
-func validateOIDCConfig(authMode AuthMode, oidcConfig *serviceApi.OIDCConfig) *common.Condition {
-	if authMode != AuthModeOIDC {
+func validateOIDCConfig(authMode cluster.AuthenticationMode, oidcConfig *serviceApi.OIDCConfig) *common.Condition {
+	if authMode != cluster.AuthModeOIDC {
 		return nil
 	}
 
@@ -407,8 +393,8 @@ func validateOIDCConfig(authMode AuthMode, oidcConfig *serviceApi.OIDCConfig) *c
 	return nil
 }
 
-func checkAuthModeNone(authMode AuthMode) *common.Condition {
-	if authMode == AuthModeNone {
+func checkAuthModeNone(authMode cluster.AuthenticationMode) *common.Condition {
+	if authMode == cluster.AuthModeNone {
 		return &common.Condition{
 			Type:    status.ConditionTypeReady,
 			Status:  metav1.ConditionFalse,
@@ -420,7 +406,7 @@ func checkAuthModeNone(authMode AuthMode) *common.Condition {
 }
 
 // getOrGenerateSecrets retrieves existing secrets or generates new ones for OAuth2 proxy.
-func getOrGenerateSecrets(ctx context.Context, rr *odhtypes.ReconciliationRequest, authMode AuthMode) (string, string, error) {
+func getOrGenerateSecrets(ctx context.Context, rr *odhtypes.ReconciliationRequest, authMode cluster.AuthenticationMode) (string, string, error) {
 	existingSecret := &corev1.Secret{}
 	secretErr := rr.Client.Get(ctx, types.NamespacedName{
 		Name:      KubeAuthProxySecretsName,
@@ -443,7 +429,7 @@ func getOrGenerateSecrets(ctx context.Context, rr *odhtypes.ReconciliationReques
 	}
 
 	var clientSecretValue string
-	if authMode == AuthModeIntegratedOAuth {
+	if authMode == cluster.AuthModeIntegratedOAuth {
 		clientSecretGen, err := cluster.NewSecret("client-secret", "random", ClientSecretLength)
 		if err != nil {
 			return "", "", fmt.Errorf("failed to generate client secret: %w", err)
