+++ internal/controller/services/gateway/resources/envoyfilter-authn.tmpl.yaml
