+++ internal/controller/services/gateway/resources/kube-auth-proxy-httproute.tmpl.yaml
@@ -1,0 +1,23 @@
+apiVersion: gateway.networking.k8s.io/v1
+kind: HTTPRoute
+metadata:
+  name: {{.OAuthCallbackRouteName}}
+  namespace: {{.GatewayNamespace}}
+spec:
+  hostnames:
+    - {{.GatewayHostname}}
+  parentRefs:
+    - name: {{.GatewayName}}
+      namespace: {{.GatewayNamespace}}
+  rules:
+    - matches:
+        - path:
+            type: PathPrefix
+            value: {{.AuthProxyOAuth2Path}}
+      backendRefs:
+        - kind: Service
+          name: {{.KubeAuthProxyServiceName}}
+          namespace: {{.GatewayNamespace}}
+          port: {{.GatewayHTTPSPort}}
+          group: ''
+          weight: 1
