+++ internal/controller/services/gateway/resources/kube-auth-proxy-oauth-deployment.tmpl.yaml
@@ -1,0 +1,101 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: {{.KubeAuthProxyServiceName}}
+  namespace: {{.GatewayNamespace}}
+  labels:
+    app: {{.KubeAuthProxyServiceName}}
+spec:
+  replicas: 2
+  selector:
+    matchLabels:
+      app: {{.KubeAuthProxyServiceName}}
+  template:
+    metadata:
+      labels:
+        app: {{.KubeAuthProxyServiceName}}
+      annotations:
+        opendatahub.io/secret-hash: "{{.AuthConfigHash}}"
+    spec:
+      securityContext:
+        runAsNonRoot: true
+        seccompProfile:
+          type: RuntimeDefault
+      volumes:
+        - name: {{.TLSCertsVolumeName}}
+          secret:
+            secretName: {{.KubeAuthProxyTLSName}}
+        - name: tmp
+          emptyDir:
+            medium: Memory
+            sizeLimit: 10Mi
+      containers:
+        - name: {{.KubeAuthProxyServiceName}}
+          image: {{.KubeAuthProxyImage}}
+          env:
+            - name: OAUTH2_PROXY_CLIENT_ID
+              valueFrom:
+                secretKeyRef:
+                  name: {{.KubeAuthProxySecretsName}}
+                  key: OAUTH2_PROXY_CLIENT_ID
+            - name: OAUTH2_PROXY_CLIENT_SECRET
+              valueFrom:
+                secretKeyRef:
+                  name: {{.KubeAuthProxySecretsName}}
+                  key: OAUTH2_PROXY_CLIENT_SECRET
+            - name: OAUTH2_PROXY_COOKIE_SECRET
+              valueFrom:
+                secretKeyRef:
+                  name: {{.KubeAuthProxySecretsName}}
+                  key: OAUTH2_PROXY_COOKIE_SECRET
+            - name: PROXY_MODE
+              value: auth
+          ports:
+            - name: http
+              containerPort: {{.AuthProxyHTTPPort}}
+            - name: https
+              containerPort: {{.GatewayHTTPSPort}}
+            - name: metrics
+              containerPort: {{.AuthProxyMetricsPort}}
+          volumeMounts:
+            - name: {{.TLSCertsVolumeName}}
+              readOnly: true
+              mountPath: {{.TLSCertsMountPath}}
+            - name: tmp
+              mountPath: /tmp
+          resources:
+            requests:
+              cpu: "10m"
+              memory: "32Mi"
+            limits:
+              cpu: "50m"
+              memory: "64Mi"
+          securityContext:
+            readOnlyRootFilesystem: true
+            allowPrivilegeEscalation: false
+            capabilities:
+              drop:
+                - ALL
+          args:
+            - "--http-address=0.0.0.0:{{.AuthProxyHTTPPort}}"
+            - "--https-address=0.0.0.0:{{.GatewayHTTPSPort}}"
+            - "--metrics-address=0.0.0.0:{{.AuthProxyMetricsPort}}"
+            - "--email-domain=*"
+            - "--upstream=static://200"
+            - "--skip-provider-button"
+            - "--skip-jwt-bearer-tokens=true"
+            - "--pass-access-token=true"
+            - "--set-xauthrequest=true"
+            - "--redirect-url={{.RedirectURL}}"
+            - "--tls-cert-file={{.TLSCertsMountPath}}/tls.crt"
+            - "--tls-key-file={{.TLSCertsMountPath}}/tls.key"
+            - "--use-system-trust-store=true"
+            - "--cookie-expire={{.CookieExpire}}"
+            - "--cookie-refresh={{.CookieRefresh}}"
+            - "--cookie-secure=true"
+            - "--cookie-httponly=true"
+            - "--cookie-samesite=lax"
+            - "--cookie-name={{.AuthProxyCookieName}}"
+            - "--cookie-domain={{.GatewayHostname}}"
+            - "--provider=openshift"
+            - "--scope=user:full"
